@@ -4,7 +4,7 @@ description = "A set of easy-to-use utils that will come in handy in any Compute
 license = { text = "MIT" }
 version = "0.26.0rc7"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 authors = [
     { name = "Piotr Skalski", email = "piotr.skalski92@gmail.com" }
 ]
@@ -75,7 +75,7 @@ docs = [
     "mkdocstrings-python>=1.10.9",
     "mike>=2.0.0",
     "mkdocs-jupyter>=0.24.3",
-    "mkdocs-git-committers-plugin-2>=2.4.1; python_version >= '3.8' and python_version < '4'",
+    "mkdocs-git-committers-plugin-2>=2.4.1; python_version >= '3.9' and python_version < '4'",
     "mkdocs-git-revision-date-localized-plugin>=1.2.4"
 ]
 build = [
@@ -93,7 +93,7 @@ check = true
 imports = ["cv2", "supervision"]
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 
 # Exclude a variety of commonly ignored directories.
 exclude = [
@@ -219,6 +219,9 @@ skip = "*.ipynb"
 count = true
 quiet-level = 3
 ignore-words-list = "STrack,sTrack,strack"
+
+[tool.pyupgrade]
+py39-plus = true
 
 [tool.setuptools]
 include-package-data = false
