@@ -1,7 +1,7 @@
 import argparse
 import json
 import os
-from typing import Any, Optional, Tuple
+from typing import Any, Optional
 
 import cv2
 import numpy as np
@@ -19,7 +19,7 @@ COLORS = sv.ColorPalette.DEFAULT
 WINDOW_NAME = "Draw Zones"
 POLYGONS = [[]]
 
-current_mouse_position: Optional[Tuple[int, int]] = None
+current_mouse_position: Optional[tuple[int, int]] = None
 
 
 def resolve_source(source_path: str) -> Optional[np.ndarray]:
