@@ -2,7 +2,7 @@ import functools
 import inspect
 import os
 import warnings
-from typing import Any, Callable, Generic, Optional, Set, TypeVar
+from typing import Any, Callable, Generic, Optional, TypeVar
 
 
 class SupervisionWarnings(Warning):
@@ -161,7 +161,7 @@ class classproperty(Generic[T]):
         return self.fget(owner_cls)
 
 
-def get_instance_variables(instance: Any, include_properties=False) -> Set[str]:
+def get_instance_variables(instance: Any, include_properties=False) -> set[str]:
     """
     Get the public variables of a class instance.
 
@@ -179,22 +179,18 @@ def get_instance_variables(instance: Any, include_properties=False) -> Set[str]:
     if isinstance(instance, type):
         raise ValueError("Only class instances are supported, not classes.")
 
-    fields = set(
-        (
-            name
-            for name, val in inspect.getmembers(instance)
-            if not callable(val) and not name.startswith("_")
-        )
-    )
+    fields = {
+        name
+        for name, val in inspect.getmembers(instance)
+        if not callable(val) and not name.startswith("_")
+    }
 
     if not include_properties:
-        properties = set(
-            (
-                name
-                for name, val in inspect.getmembers(instance.__class__)
-                if isinstance(val, property)
-            )
-        )
+        properties = {
+            name
+            for name, val in inspect.getmembers(instance.__class__)
+            if isinstance(val, property)
+        }
         fields -= properties
 
     return fields
