@@ -2,8 +2,9 @@ from __future__ import annotations
 
 import time
 from collections import deque
+from collections.abc import Generator
 from dataclasses import dataclass
-from typing import Callable, Generator, Optional, Tuple
+from typing import Callable
 
 import cv2
 import numpy as np
@@ -40,7 +41,7 @@ class VideoInfo:
     width: int
     height: int
     fps: int
-    total_frames: Optional[int] = None
+    total_frames: int | None = None
 
     @classmethod
     def from_video_path(cls, video_path: str) -> VideoInfo:
@@ -56,7 +57,7 @@ class VideoInfo:
         return VideoInfo(width, height, fps, total_frames)
 
     @property
-    def resolution_wh(self) -> Tuple[int, int]:
+    def resolution_wh(self) -> tuple[int, int]:
         return self.width, self.height
 
 
@@ -118,7 +119,7 @@ class VideoSink:
 
 
 def _validate_and_setup_video(
-    source_path: str, start: int, end: Optional[int], iterative_seek: bool = False
+    source_path: str, start: int, end: int | None, iterative_seek: bool = False
 ):
     video = cv2.VideoCapture(source_path)
     if not video.isOpened():
@@ -145,9 +146,9 @@ def get_video_frames_generator(
     source_path: str,
     stride: int = 1,
     start: int = 0,
-    end: Optional[int] = None,
+    end: int | None = None,
     iterative_seek: bool = False,
-) -> Generator[np.ndarray, None, None]:
+) -> Generator[np.ndarray]:
     """
     Get a generator that yields the frames of the video.
 
@@ -196,7 +197,7 @@ def process_video(
     source_path: str,
     target_path: str,
     callback: Callable[[np.ndarray, int], np.ndarray],
-    max_frames: Optional[int] = None,
+    max_frames: int | None = None,
     show_progress: bool = False,
     progress_message: str = "Processing video",
 ) -> None:
