@@ -1,5 +1,5 @@
 from contextlib import ExitStack as DoesNotRaise
-from typing import Dict, List, Optional, Tuple, TypeVar
+from typing import Optional, TypeVar
 
 import numpy as np
 import numpy.typing as npt
@@ -74,11 +74,11 @@ T = TypeVar("T")
     ],
 )
 def test_train_test_split(
-    data: List[T],
+    data: list[T],
     train_ratio: float,
     random_state: int,
     shuffle: bool,
-    expected_result: Optional[Tuple[List[T], List[T]]],
+    expected_result: Optional[tuple[list[T], list[T]]],
     exception: Exception,
 ) -> None:
     with exception:
@@ -118,7 +118,7 @@ def test_train_test_split(
     ],
 )
 def test_merge_class_maps(
-    class_lists: List[List[str]], expected_result: List[str], exception: Exception
+    class_lists: list[list[str]], expected_result: list[str], exception: Exception
 ) -> None:
     with exception:
         result = merge_class_lists(class_lists=class_lists)
@@ -163,9 +163,9 @@ def test_merge_class_maps(
     ],
 )
 def test_build_class_index_mapping(
-    source_classes: List[str],
-    target_classes: List[str],
-    expected_result: Optional[Dict[int, int]],
+    source_classes: list[str],
+    target_classes: list[str],
+    expected_result: Optional[dict[int, int]],
     exception: Exception,
 ) -> None:
     with exception:
@@ -223,7 +223,7 @@ def test_build_class_index_mapping(
     ],
 )
 def test_map_detections_class_id(
-    source_to_target_mapping: Dict[int, int],
+    source_to_target_mapping: dict[int, int],
     detections: Detections,
     expected_result: Optional[Detections],
     exception: Exception,
@@ -287,7 +287,7 @@ def test_map_detections_class_id(
     ],
 )
 def test_mask_to_rle(
-    mask: npt.NDArray[np.bool_], expected_rle: List[int], exception: Exception
+    mask: npt.NDArray[np.bool_], expected_rle: list[int], exception: Exception
 ) -> None:
     with exception:
         result = mask_to_rle(mask=mask)
@@ -354,7 +354,7 @@ def test_mask_to_rle(
 )
 def test_rle_to_mask(
     rle: npt.NDArray[np.int_],
-    resolution_wh: Tuple[int, int],
+    resolution_wh: tuple[int, int],
     expected_mask: npt.NDArray[np.bool_],
     exception: Exception,
 ) -> None:
