@@ -1,5 +1,5 @@
 from contextlib import ExitStack as DoesNotRaise
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Optional
 
 import numpy as np
 import numpy.typing as npt
@@ -66,7 +66,7 @@ TEST_MASK[:, 300:351, 200:251] = True
 )
 def test_clip_boxes(
     xyxy: np.ndarray,
-    resolution_wh: Tuple[int, int],
+    resolution_wh: tuple[int, int],
     expected_result: np.ndarray,
 ) -> None:
     result = clip_boxes(xyxy=xyxy, resolution_wh=resolution_wh)
@@ -152,10 +152,10 @@ def test_clip_boxes(
     ],
 )
 def test_filter_polygons_by_area(
-    polygons: List[np.ndarray],
+    polygons: list[np.ndarray],
     min_area: Optional[float],
     max_area: Optional[float],
-    expected_result: List[np.ndarray],
+    expected_result: list[np.ndarray],
     exception: Exception,
 ) -> None:
     with exception:
@@ -372,7 +372,7 @@ def test_filter_polygons_by_area(
 )
 def test_process_roboflow_result(
     roboflow_result: dict,
-    expected_result: Tuple[
+    expected_result: tuple[
         np.ndarray, np.ndarray, np.ndarray, Optional[np.ndarray], np.ndarray
     ],
     exception: Exception,
@@ -698,7 +698,7 @@ def test_move_boxes(
 def test_move_masks(
     masks: np.ndarray,
     offset: np.ndarray,
-    resolution_wh: Tuple[int, int],
+    resolution_wh: tuple[int, int],
     expected_result: np.ndarray,
     exception: Exception,
 ) -> None:
@@ -1032,8 +1032,8 @@ def test_calculate_masks_centroids(
     ],
 )
 def test_merge_data(
-    data_list: List[Dict[str, Any]],
-    expected_result: Optional[Dict[str, Any]],
+    data_list: list[dict[str, Any]],
+    expected_result: Optional[dict[str, Any]],
     exception: Exception,
 ):
     with exception:
@@ -1207,9 +1207,9 @@ def test_merge_data(
     ],
 )
 def test_get_data_item(
-    data: Dict[str, Any],
+    data: dict[str, Any],
     index: Any,
-    expected_result: Optional[Dict[str, Any]],
+    expected_result: Optional[dict[str, Any]],
     exception: Exception,
 ):
     with exception:
