@@ -1,6 +1,6 @@
 import os
 from contextlib import ExitStack as DoesNotRaise
-from typing import List, Optional
+from typing import Optional
 
 import pytest
 
@@ -56,7 +56,7 @@ def setup_and_teardown_files():
 def test_read_txt_file(
     file_name: str,
     skip_empty: bool,
-    expected_result: Optional[List[str]],
+    expected_result: Optional[list[str]],
     exception: Exception,
 ):
     with exception:
